% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_whitespace.R
\name{collapse_whitespace}
\alias{collapse_whitespace}
\title{Collapse whitespace}
\usage{
collapse_whitespace(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string
}
\description{
Replaces each unquoted run of whitespace characters with a
  single space
}
