% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alias.R
\name{extract_alias}
\alias{extract_alias}
\title{Extract column alias from SQL expression}
\usage{
extract_alias(expr)
}
\arguments{
\item{expr}{a character string containing a SQL expression which might have a
column alias assignment at the end}
}
\value{
a character string containing the inputed SQL expression with the
  column alias assignment removed (if it was present) and with the assigned
  alias as its name
}
\description{
Extracts the column alias from an expression used in the SELECT
  list of a SQL query
}
\examples{
expr <- "round(AVG(arr_delay)) AS avg_delay"
extract_alias(expr)
}
