% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_query.R
\name{lex_query}
\alias{lex_query}
\title{SQL query lexer}
\usage{
lex_query(query)
}
\arguments{
\item{query}{a SQL SELECT statement}
}
\value{
A list object with named elements representing the clauses of the
  query
}
\description{
Splits a SQL SELECT statement into clauses and splits
  comma-separated column lists within the clauses. Finer-grained lexing is
  performed by \code{\link{parse_query}}.
}
\examples{
query <- "SELECT origin, dest,
    COUNT(*) AS num_flts,
    round(AVG(distance)) AS dist,
    round(AVG(arr_delay)) AS avg_delay
  FROM flights
  WHERE distance BETWEEN 200 AND 300
  GROUP BY origin, dest
  HAVING num_flts > 5000
  ORDER BY num_flts DESC, avg_delay DESC
  LIMIT 100;"
lex_query(query)
}
\seealso{
\code{\link{parse_query}}
}
