% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query.R
\name{parse_query}
\alias{parse_query}
\title{SQL query parser}
\usage{
parse_query(query)
}
\arguments{
\item{query}{a character string containing a SQL SELECT statement}
}
\value{
A list object with named elements representing the clauses of the
  query
}
\description{
Parses a SQL SELECT statement and returns a parse tree
}
\examples{
query <- "SELECT origin, dest,
    COUNT(*) AS num_flts,
    round(AVG(distance)) AS dist,
    round(AVG(arr_delay)) AS avg_delay
  FROM flights
  WHERE distance BETWEEN 200 AND 300
  GROUP BY origin, dest
  HAVING num_flts > 5000
  ORDER BY num_flts DESC, avg_delay DESC
  LIMIT 100;"
parse_query(query)
}
