% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_sql.R
\name{squish_sql}
\alias{squish_sql}
\title{Squish a SQL query or SQL expression}
\usage{
squish_sql(x)
}
\arguments{
\item{x}{a character string containing a SQL query or expression}
}
\value{
a character string containing the squished query or expression with
  comments removed
}
\description{
Replaces every unquoted run of whitespace characters with a
  single space and removes all line comments (\code{--}) and block comments
  (\code{/* */}). Whitespace and comment marks within quotes are not
  modified.
}
