% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unqualify.R
\name{unqualify_query}
\alias{unqualify_query}
\title{Remove prefixes from column references in a SQL query}
\usage{
unqualify_query(tree, prefixes)
}
\arguments{
\item{tree}{a list returned by \code{\link{squish_sql}} containing named
elements representing the clauses of a SQL \code{SELECT} statement}

\item{prefixes}{a character vector containing one or more table names or
table aliases}
}
\value{
a list the same as \code{tree} but with all column references in the
  \code{SELECT}, \code{WHERE}, \code{GROUP BY}, \code{HAVING}, and
  \code{ORDER BY} clauses unqualified. The \code{FROM} clause is unmodified
}
\description{
Unqualifies all column references in the clauses of a SQL
  \code{SELECT} statement that begin with any of the specified prefixes
  followed by a dot
}
